==> analysis/R/Scripts/graveyard/01_data_properties.R <==
######################################

# This script:
# - Takes a dataset and summarises the variables using:
#   * skimr::skim(),
#   * class(),
#   * walk2()
# - It saves the output to a .txt file
# - The script should only be run via an action in the project.yaml only
# - The script must be accompanied by two arguments
# - The first is the dataset, saved as an .rds file, that is to be summarised
# - The second in the directory where the txt output will be saved

######################################


# Preliminaries ----

## Import libraries
library('tidyverse')
source(here::here("lib", "redaction_functions.R"))

if(Sys.getenv("OPENSAFELY_BACKEND") %in% c("", "expectations")){
  for(backend in c("tpp", "emis")){
    
    ## Import command-line arguments
    rds_file <- "output/data/data_modelling.rds"
    output_dir <- paste("output/", {backend}, "/data_properties", sep = "")
    
    stopifnot("must pass an .rds file" = fs::path_ext(rds_file) == "rds")
    
    ## Base file name
    filenamebase <- fs::path_ext_remove(fs::path_file(rds_file))
    
    ## Import processed data
    data <- readr::read_rds(here::here(rds_file))
    
    # Output summary .txt
    options(width = 200) # set output width for capture.output
    dir.create(here::here(output_dir), showWarnings = FALSE, recursive = TRUE)
    
    # Summaries ----
    
    ## High-level variable overview
    capture.output(
      skimr::skim_without_charts(data),
      file = here::here(output_dir, paste0(filenamebase, "_skim", ".txt")),
      split = FALSE
    )
    
    ## List of column types
    capture.output(
      lapply(data, class),
      file = here::here(output_dir, paste0(filenamebase, "_coltypes", ".txt"))
    )
    
    
    ## Tabulated data
    
    ### Delete file if it exists
    if(file.exists(here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")))){
      file.remove(here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")))
    }
    
    ### Categorical and logical
    sumtabs_cat <-
      data %>%
      select(-ends_with("_id")) %>%
      select(where(is.character), where(is.logical), where(is.factor)) %>%
      map(redacted_summary_cat) %>%
      enframe()
    
    capture.output(
      walk2(sumtabs_cat$value, sumtabs_cat$name, print_cat),
      file = here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")),
      append=FALSE
    )
    
    ### Numeric
    sumtabs_num <-
      data %>%
      select(-ends_with("_id")) %>%
      select(where(~ {!is.logical(.x) & is.numeric(.x) & !is.Date(.x)})) %>%
      map(redacted_summary_num) %>%
      enframe()
    
    capture.output(
      walk2(sumtabs_num$value, sumtabs_num$name, print_num),
      file = here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")),
      append=TRUE
    )
    
    ### Dates
    sumtabs_date <-
      data %>%
      select(-ends_with("_id")) %>%
      select(where(is.Date)) %>%
      map(redacted_summary_date) %>%
      enframe()
    
    capture.output(
      walk2(sumtabs_date$value, sumtabs_date$name, print_num),
      file = here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")),
      append=TRUE
    )
    
  }
} else {
  
  ## Import command-line arguments
  rds_file <- "output/data/data_modelling.rds"
  output_dir <- paste("output/", Sys.getenv('OPENSAFELY_BACKEND'), "/data_properties", sep = "")
  
  stopifnot("must pass an .rds file" = fs::path_ext(rds_file) == "rds")
  
  ## Base file name
  filenamebase <- fs::path_ext_remove(fs::path_file(rds_file))
  
  ## Import processed data
  data <- readr::read_rds(here::here(rds_file))
  
  # Output summary .txt
  options(width = 200) # set output width for capture.output
  dir.create(here::here(output_dir), showWarnings = FALSE, recursive = TRUE)
  
  # Summaries ----
  
  ## High-level variable overview
  capture.output(
    skimr::skim_without_charts(data),
    file = here::here(output_dir, paste0(filenamebase, "_skim", ".txt")),
    split = FALSE
  )
  
  ## List of column types
  capture.output(
    lapply(data, class),
    file = here::here(output_dir, paste0(filenamebase, "_coltypes", ".txt"))
  )
  
  
  ## Tabulated data
  
  ### Delete file if it exists
  if(file.exists(here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")))){
    file.remove(here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")))
  }
  
  ### Categorical and logical
  sumtabs_cat <-
    data %>%
    select(-ends_with("_id")) %>%
    select(where(is.character), where(is.logical), where(is.factor)) %>%
    map(redacted_summary_cat) %>%
    enframe()
  
  capture.output(
    walk2(sumtabs_cat$value, sumtabs_cat$name, print_cat),
    file = here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")),
    append=FALSE
  )
  
  ### Numeric
  sumtabs_num <-
    data %>%
    select(-ends_with("_id")) %>%
    select(where(~ {!is.logical(.x) & is.numeric(.x) & !is.Date(.x)})) %>%
    map(redacted_summary_num) %>%
    enframe()
  
  capture.output(
    walk2(sumtabs_num$value, sumtabs_num$name, print_num),
    file = here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")),
    append=TRUE
  )
  
  ### Dates
  sumtabs_date <-
    data %>%
    select(-ends_with("_id")) %>%
    select(where(is.Date)) %>%
    map(redacted_summary_date) %>%
    enframe()
  
  capture.output(
    walk2(sumtabs_date$value, sumtabs_date$name, print_num),
    file = here::here(output_dir, paste0(filenamebase, "_tabulate", ".txt")),
    append=TRUE
  )
  
}
